+++ screens/(setting)/NoticeScreen.tsx
@@ -1,37 +1,24 @@
-import { View } from 'react-native';
-import { supabase } from 'lib/supabase';
-import { useEffect, useState } from 'react';
-import DropDown from '@components/page/setting/DropDown';
-
-interface NoticeType {
-  id: number;
-  title: string;
-  content: string;
-  created_at: string;
-}
+import { StyleSheet, View } from 'react-native';
+import { Suspense } from 'react';
+import Loading from '@components/common/Loading';
+import Notice from '@components/page/notice/Notice';
+import SettingHeader from '@components/page/setting/SettingHeader';
 
 const NoticeScreen = () => {
-  const [notices, setNotices] = useState<NoticeType[]>([]);
-
-  useEffect(() => {
-    getNotice();
-  }, []);
-
-  const getNotice = async () => {
-    const { data } = await supabase
-      .from('notices')
-      .select()
-      .returns<NoticeType[]>();
-    if (data) setNotices(data);
-  };
-
   return (
-    <View>
-      {notices.map((notice, index) => (
-        <DropDown key={index} item={notice} index={index} />
-      ))}
+    <View style={s.container}>
+      <SettingHeader title="공지사항" />
+      <Suspense fallback={<Loading />}>
+        <Notice />
+      </Suspense>
     </View>
   );
 };
 
 export default NoticeScreen;
+
+const s = StyleSheet.create({
+  container: {
+    flex: 1,
+  },
+});
